requirements.txt
crewai>=0.130.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
openai>=1.3.0


